Created 'EventCreate' component and added functionality
This commit introduces a new component called 'EventCreate' for creating events. It includes the HTML form design, relevant TypeScript logic, and tests. Alongside this, important updates were made to the 'EventListComponent' and 'AppModule' to implement and route the new component. Additionally, a dialog service was added to allow the creation of new events through a pop-up form.